refactored types
