+++ src/components/product/types.ts
